pyproject.toml
[project]
name = "brawlifics"
description = "Brawl Stars + Scientifics = Educational Game"
authors = [
  {name = "Oriol Rius", email = "oriol@joor.net"},
  {name = "Néstor López", email = "nestor@nstlopez.com"},
  ]
version = "0.9.0"
readme = "README.md"
requires-python = "~=3.13"
dependencies = [
  "fastapi>=0.109.0,<0.110",
  "uvicorn>=0.27.0,<0.28",
  "paho-mqtt>=2.1.0",
  "python-dotenv>=1.0.0,<2",
  "httpx>=0.26.0,<0.27",
  "pyyaml>=6.0.2,<7",
  "pillow>=11.0.0,<12",
  "python-multipart>=0.0.19,<0.0.20",
  "tomli>=2.2.1",
  "ruff>=0.8.6",
]

[dependency-groups]
dev = [
  "pytest>=7.4.4,<8",
  "pytest-asyncio>=0.23.3,<0.24",
  "isort>=5.13.2,<6",
]

[tool.hatch.build.targets.sdist]
include = ["brawlifics"]

[tool.hatch.build.targets.wheel]
include = ["brawlifics"]

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.pytest.ini_options]
markers = [
  "integration: marks tests as integration tests",
]
